SQLite: add incremental vacuuming
